i need sleep
currently need to fix the sql to make the add movie work and be able to test it. Also need to look at how to display the information to the rent movie window and how we are going to pass that information back into the db. still need to add some more stored procedures as well